Refine data storage load/save and add unit test framework
The data storage framework has been refined to improve data load/save,
to allow historical data to be loaded and tested against, and to support
the management of the 14 days of storage requirement.

The unit test scaffold has been put in place. The unit tests themselves
are still to be added.