Align API with GApple API 1.3.1
Aligns the external API with the Google/Apple API version 1.3.1 as
detailed in the Exposure Notification API Documentation:

https://www.google.com/covid19/exposurenotifications/pdfs/Android-Exposure-Notification-API-documentation-v1.3.1.pdf

This also includes the protobuf generated code needed for reading in the
Diagnosis keys obtained from the server.

As yet, although the API exists, the functionality hasn't yet been
hooked up to it.