Restructure projec to include daemon
Adds a new target an various files needed for running the contracd
daemon process.

Currently the daemon isn't functional; the code will need to be moved
from the app into the daemon and the daemon will need to be given a dbus
interface for the app to talk to it through.