Move contrac code into the daemon
All of the core contrac functionality should happen within the daemon,
rather than the front end application. This change moves all of the core
classes from the app to the daemon.

It also hooks up the various time-based signals and slots to allow
everything to be accessed through the single interface.