Add encrypted metadata to beacons
Adds four bytes of encrypted metadata to the advertising beacons. The
data can only be decrypted if the tracing key is known.

This also fixes multiple issues with the exposure notification data
aggregation routines:

1. Correctly calculates attenuation scores based on the specification.
See here for more details:
https://developers.google.com/android/exposure-notifications/ble-attenuation-overview

2. Aggregates based on time more effectively using ctintervals (every 2
seconds) rather than intervals (every 10 minutes).

3. Sets up default values for the configuration structure (previously
values were left uninitialised on creation).

4. Items are no longer removed from the bloomfilter probable contact
list when discovered. This allows multiple beacons with the same RPI to
be incorporated into the calculation, rather than being skipped, since
the RPIs may fall in different ctintervals, or have different RSSI
values, etc.

5. The aggregateExposureData() function no longer assumes the incomming
RPIs are ordered by time (it performs an order check and sort before
applying the aggregation algorithm).

6. Output times are rounded more correctly to the nearest minute or five
minutes and with an upper limit as defined by the spec.

The change adds in unit tests for checking the aggregation algorithm and
metadata validity, and improves the contact matching unit test to take
into account the time interval as well as the day.