Adds config download from server
The ExposureConfiguration structure needs to be filled out with values
in order for the risk calculations to be applied correctly.

This change adds a step which downloads the configuration values from
the download server in proto format, extracts the relevant values and
returns them in an ExposureConfiguration structure which can be passed
to the Contrac daemon.