Fix download frequency timer
Downloads are only allowed once per day, but the calculation used to
decide whether to permit a fresh download was incorrect.

This change fixes it so that a new download will be allowed as soon as
the clock turns midnight (local time) after the last download occurred.

This also removes the uploadAvailable property, since it wasn't being
used.