Add combined risk score calculation
The Corona-Warn-App does some post-processing to refine the risk level
calculates by the GApple API, as explained here:

https://github.com/corona-warn-app/cwa-documentation/blob/master/solution_architecture.md#risk-score-calculation

In short, this multiplies each of the attenuation durations by a weight,
adds them together along with a constant, multiplies the result by the
GApple risk score, and divides by a normalizing factor.

The weights, constant and factor are supplied through the configuration
file downloaded from the servers.

This change stores the configuration values and provides methods for
calculating the combined risk score.