Update to work with OpenSSL 1.1.1
The OpenSSL library interface has changed in the upgrade from 1.0.2o to
1.1.1.h. The following affect Contrac:

1. The EVP_MD_CTX_create() and EVP_MD_CTX_destroy() functions have been
renamed to EVP_MD_CTX_new() and EVP_MD_CTX_free() respectively.

2. EVP_PKEY_HKDF has been added to allow derivation of HKDF keys.
Previously we used the OpenSSL HMAC interface to derive our own, but
since the interface needed has changed, we now use EVP_PKEY_HKDF
directly.

The new (OpenSSL 1.1.1) implementations are used if the OPENSSL_GE_1_1_1
directive is defined. The switch for this is in version.pri.

The plan is that the OpenSSL 1.0.2 code can eventually be removed
entirely.