Move risk calculation into a separate thread
The risk calculation triggered when ProvideDiagnosticKeys is called can
take a long time, so it's better to run it on a separate thread in order
to allow other tasks (such as the beacon send/receive) to continue
running while the calculation takes place.

This moves the calculation into its own thread. It also changes the API
slightly so that the provideDiagnosisKeys() dbus call no longer blocks,
but instead notifies completion using the actionExposureStateUpdated()
signal.

This also better matches the GAEN API.