Expose processing time through dbus interface
Exposes the date/time of the last diagnosis key processing for a given
token through the dbus interface.

This allows the risk status to be updated even if the processing
completed in the background while the app was shut.

The change also re-works the background processing thread to reduce the
need to access data from the main thread while the processing is
running in order to reduce mutex use.